+++ components/Tabs/tests/unit/TabBarView/MDCTabBarViewTests.m
@@ -273,6 +273,12 @@ static UIImage *fakeImage(CGSize size) {
   XCTAssertEqual(self.tabBarView.barTintColor, self.tabBarView.backgroundColor);
 }
 
+- (void)testDefaultRippleColor {
+  // Then
+  XCTAssertEqualObjects(self.tabBarView.rippleColor, [[UIColor alloc] initWithWhite:0
+                                                                              alpha:(CGFloat)0.16]);
+}
+
 - (void)testImageTintColorForStateFallsBackToNormalState {
   // Given
   [self.tabBarView setImageTintColor:nil forState:UIControlStateNormal];
@@ -653,7 +659,7 @@ static UIImage *fakeImage(CGSize size) {
   XCTAssertEqualWithAccuracy(size.height, kMinHeight, 0.001);
 }
 
-- (void)testSizeThatFitsDoesntShrinkToFitContent {
+- (void)testSizeThatFitsShrinksToFitContentVerticallyButNotHorizontally {
   // Given
   self.tabBarView.items = @[ self.itemA ];
   CGSize intrinsicSize = self.tabBarView.intrinsicContentSize;
@@ -664,7 +670,7 @@ static UIImage *fakeImage(CGSize size) {
 
   // Then
   XCTAssertEqualWithAccuracy(size.width, biggerSize.width, 0.001);
-  XCTAssertEqualWithAccuracy(size.height, biggerSize.height, 0.001);
+  XCTAssertEqualWithAccuracy(size.height, intrinsicSize.height, 0.001);
 }
 
 @end